Split out thread milling generator into its own file
